Memory optimization on imports
